[Refactor] MyPageLevel 위치 변경
