+++ feature/mypage/src/main/java/team/ppac/mypage/component/MyPageProgressBar.kt
@@ -22,7 +22,7 @@ import androidx.compose.ui.unit.dp
 import team.ppac.designsystem.FarmemeTheme
 import team.ppac.designsystem.foundation.FarmemeIcon
 import team.ppac.designsystem.foundation.FarmemeRadius
-import team.ppac.mypage.MyPageLevel
+import team.ppac.mypage.model.MyPageLevel
 
 @Composable
 internal fun MyPageProgressBar(
